Update to TDLib 1.7.0

+++ src/types/option_value.rs
@@ -234,7 +234,7 @@ pub struct OptionValueInteger {
   #[serde(rename(serialize = "@type", deserialize = "@type"))]
   td_name: String,
   /// The value of the option
-  value: i64,
+  value: isize,
   
 }
 
@@ -256,7 +256,7 @@ impl OptionValueInteger {
     RTDOptionValueIntegerBuilder { inner }
   }
 
-  pub fn value(&self) -> i64 { self.value }
+  pub fn value(&self) -> isize { self.value }
 
 }
 
@@ -269,7 +269,7 @@ impl RTDOptionValueIntegerBuilder {
   pub fn build(&self) -> OptionValueInteger { self.inner.clone() }
 
    
-  pub fn value(&mut self, value: i64) -> &mut Self {
+  pub fn value(&mut self, value: isize) -> &mut Self {
     self.inner.value = value;
     self
   }
